Change namespace `WPCLI` to `CLI`
