+++ src/CLI/Commands/ClearCommand.php
@@ -1,18 +1,16 @@
 <?php
 
-namespace Hirasso\WP\ThumbhashPlaceholders\WPCLI\Commands;
+namespace Hirasso\WP\ThumbhashPlaceholders\CLI\Commands;
 
-use Hirasso\WP\ThumbhashPlaceholders\ImageDownloader;
 use Hirasso\WP\ThumbhashPlaceholders\Plugin;
-use Hirasso\WP\ThumbhashPlaceholders\WPCLI\InputValidator;
-use Hirasso\WP\ThumbhashPlaceholders\WPCLI\Utils;
+use Hirasso\WP\ThumbhashPlaceholders\CLI\InputValidator;
+use Hirasso\WP\ThumbhashPlaceholders\CLI\Utils;
 use Snicco\Component\BetterWPCLI\Command;
 use Snicco\Component\BetterWPCLI\Input\Input;
 use Snicco\Component\BetterWPCLI\Output\Output;
 use Snicco\Component\BetterWPCLI\Style\SniccoStyle;
 use Snicco\Component\BetterWPCLI\Style\Text;
 use Snicco\Component\BetterWPCLI\Synopsis\InputArgument;
-use Snicco\Component\BetterWPCLI\Synopsis\InputFlag;
 use Snicco\Component\BetterWPCLI\Synopsis\Synopsis;
 use WP_Query;
 
@@ -20,11 +18,11 @@ use WP_Query;
  * A WP CLI command to generate thumbhash placeholders
  * @see https://github.com/snicco/better-wp-cli
  */
-class GenerateCommand extends Command
+class ClearCommand extends Command
 {
-    protected static string $name = 'generate';
+    protected static string $name = 'clear';
 
-    protected static string $short_description = 'Generate thumbhash placeholders';
+    protected static string $short_description = 'Clear thumbhash placeholders';
 
     /**
      * Command synopsis.
@@ -37,24 +35,16 @@ class GenerateCommand extends Command
                 'Only generate placeholders for these images',
                 InputArgument::OPTIONAL | InputArgument::REPEATING
             ),
-            new InputFlag(
-                'force',
-                'Generate placeholders also for images that already have one'
-            ),
         );
     }
 
-    /**
-     * Execute the command
-     */
     public function execute(Input $input, Output $output): int
     {
         $io = new SniccoStyle($input, $output);
 
         $ids = $input->getRepeatingArgument('ids', []);
-        $force = $input->getFlag('force');
 
-        $io->title("Generating ThumbHash Placeholders");
+        $io->title("Clearing ThumbHash Placeholders");
 
         $validator = new InputValidator($io);
         if (!$validator->isNumericArray($ids, "Non-numeric ids provided")) {
@@ -66,48 +56,41 @@ class GenerateCommand extends Command
             'post_status' => 'inherit',
             'post_mime_type' => 'image',
             'posts_per_page' => -1,
-            'post__in' => $ids,
             'fields' => 'ids',
             // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_query -- this only ever runs in WP CLI
             'meta_query' => [
                 [
                     'key' => Plugin::META_KEY,
-                    'compare' => 'NOT EXISTS',
+                    'compare' => 'EXISTS',
                 ],
             ],
         ];
-
-        if ($force) {
-            unset($queryArgs['meta_query']);
+        if (!empty($ids)) {
+            $queryArgs['post__in'] = array_map('absint', array_map('trim', $ids));
         }
 
         $query = new WP_Query($queryArgs);
 
-        ImageDownloader::cleanupOldImages();
-
         if (!$query->have_posts()) {
-            $io->success("No images without placeholders found");
+            $io->success("No images with placeholders found");
             return Command::SUCCESS;
         }
-
         $count = 0;
         foreach ($query->posts as $id) {
-            $thumbhash = Plugin::generateThumbhash($id);
-            $status = match (!!$thumbhash) {
-                true => $io->colorize('generated ✔︎', Text::GREEN),
-                default => $io->colorize('failed ❌', Text::RED)
-            };
-            $output->writeln(Utils::getStatusLine(basename(wp_get_attachment_url($id)), $status));
-            if ($thumbhash) {
-                $count++;
-            }
+            delete_post_meta($id, Plugin::META_KEY);
+            $output->writeln(Utils::getStatusLine(
+                basename(wp_get_attachment_url($id)),
+                $io->colorize('cleared ✔︎', Text::GREEN)
+            ));
+            $count++;
         }
+
         $output->newLine();
 
         $io->success(match ($count) {
-            1 => "$count placeholder generated",
-            0 => "No placeholders generated",
-            default => "$count placeholders generated"
+            1 => "$count placeholder cleared",
+            0 => "No placeholders cleared",
+            default => "$count placeholders cleared"
         });
 
         return Command::SUCCESS;
