+++ src/CLI/Commands/GenerateCommand.php
@@ -1,16 +1,18 @@
 <?php
 
-namespace Hirasso\WP\ThumbhashPlaceholders\WPCLI\Commands;
+namespace Hirasso\WP\ThumbhashPlaceholders\CLI\Commands;
 
+use Hirasso\WP\ThumbhashPlaceholders\ImageDownloader;
 use Hirasso\WP\ThumbhashPlaceholders\Plugin;
-use Hirasso\WP\ThumbhashPlaceholders\WPCLI\InputValidator;
-use Hirasso\WP\ThumbhashPlaceholders\WPCLI\Utils;
+use Hirasso\WP\ThumbhashPlaceholders\CLI\InputValidator;
+use Hirasso\WP\ThumbhashPlaceholders\CLI\Utils;
 use Snicco\Component\BetterWPCLI\Command;
 use Snicco\Component\BetterWPCLI\Input\Input;
 use Snicco\Component\BetterWPCLI\Output\Output;
 use Snicco\Component\BetterWPCLI\Style\SniccoStyle;
 use Snicco\Component\BetterWPCLI\Style\Text;
 use Snicco\Component\BetterWPCLI\Synopsis\InputArgument;
+use Snicco\Component\BetterWPCLI\Synopsis\InputFlag;
 use Snicco\Component\BetterWPCLI\Synopsis\Synopsis;
 use WP_Query;
 
@@ -18,11 +20,11 @@ use WP_Query;
  * A WP CLI command to generate thumbhash placeholders
  * @see https://github.com/snicco/better-wp-cli
  */
-class ClearCommand extends Command
+class GenerateCommand extends Command
 {
-    protected static string $name = 'clear';
+    protected static string $name = 'generate';
 
-    protected static string $short_description = 'Clear thumbhash placeholders';
+    protected static string $short_description = 'Generate thumbhash placeholders';
 
     /**
      * Command synopsis.
@@ -35,16 +37,24 @@ class ClearCommand extends Command
                 'Only generate placeholders for these images',
                 InputArgument::OPTIONAL | InputArgument::REPEATING
             ),
+            new InputFlag(
+                'force',
+                'Generate placeholders also for images that already have one'
+            ),
         );
     }
 
+    /**
+     * Execute the command
+     */
     public function execute(Input $input, Output $output): int
     {
         $io = new SniccoStyle($input, $output);
 
         $ids = $input->getRepeatingArgument('ids', []);
+        $force = $input->getFlag('force');
 
-        $io->title("Clearing ThumbHash Placeholders");
+        $io->title("Generating ThumbHash Placeholders");
 
         $validator = new InputValidator($io);
         if (!$validator->isNumericArray($ids, "Non-numeric ids provided")) {
@@ -56,41 +66,48 @@ class ClearCommand extends Command
             'post_status' => 'inherit',
             'post_mime_type' => 'image',
             'posts_per_page' => -1,
+            'post__in' => $ids,
             'fields' => 'ids',
             // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_query -- this only ever runs in WP CLI
             'meta_query' => [
                 [
                     'key' => Plugin::META_KEY,
-                    'compare' => 'EXISTS',
+                    'compare' => 'NOT EXISTS',
                 ],
             ],
         ];
-        if (!empty($ids)) {
-            $queryArgs['post__in'] = array_map('absint', array_map('trim', $ids));
+
+        if ($force) {
+            unset($queryArgs['meta_query']);
         }
 
         $query = new WP_Query($queryArgs);
 
+        ImageDownloader::cleanupOldImages();
+
         if (!$query->have_posts()) {
-            $io->success("No images with placeholders found");
+            $io->success("No images without placeholders found");
             return Command::SUCCESS;
         }
+
         $count = 0;
         foreach ($query->posts as $id) {
-            delete_post_meta($id, Plugin::META_KEY);
-            $output->writeln(Utils::getStatusLine(
-                basename(wp_get_attachment_url($id)),
-                $io->colorize('cleared ✔︎', Text::GREEN)
-            ));
-            $count++;
+            $thumbhash = Plugin::generateThumbhash($id);
+            $status = match (!!$thumbhash) {
+                true => $io->colorize('generated ✔︎', Text::GREEN),
+                default => $io->colorize('failed ❌', Text::RED)
+            };
+            $output->writeln(Utils::getStatusLine(basename(wp_get_attachment_url($id)), $status));
+            if ($thumbhash) {
+                $count++;
+            }
         }
-
         $output->newLine();
 
         $io->success(match ($count) {
-            1 => "$count placeholder cleared",
-            0 => "No placeholders cleared",
-            default => "$count placeholders cleared"
+            1 => "$count placeholder generated",
+            0 => "No placeholders generated",
+            default => "$count placeholders generated"
         });
 
         return Command::SUCCESS;
